- Removes duplicate of the example charms - Adapts documentation to reflect proper example charms - Creates placeholder charm
