+++ CONTRIBUTING.md
@@ -9,20 +9,10 @@ Create and activate a virtualenv with the development requirements:
 
 ## Testing
 
-### Unit tests
+Testing for this project is done using `tox`. You can run the various tests like so:
 
-```bash
-tox -e unit
-```
-
-### Static analysis
-
-```bash
-tox -e static
-```
-
-### Linting
-
-```bash
-tox -e lint
+```shell
+tox -e lint      # code style
+tox -e static    # static analysis
+tox -e unit      # unit tests
 ```
