+++ charmcraft.yaml
@@ -6,7 +6,3 @@ bases:
     run-on:
     - name: "ubuntu"
       channel: "20.04"
-parts:
-  charm:
-    build-packages:
-      - git
