+++ requirements.txt
@@ -1,3 +1,1 @@
-git+https://github.com/canonical/operator/#egg=ops
-jsonschema
-cryptography
+ops
